improving treectrl and 2D Curve drawing
git-svn-id: file:///home/karl/SVN2/amilab.svn/code/trunk@432 52b0d089-85ac-4fb5-acf5-4e3172ea7aaf

+++ doc/generate_html.py
@@ -569,9 +569,10 @@ if __name__ == "__main__":
       # removing the files under the directory Rules
       for root, dirs, files in os.walk("Rules"):
         for name in files:
-          os.remove(os.path.join(root, name))
-      os.rmdir("Rules")
-    os.mkdir("Rules")
+          if (name[-5:]==".html"):
+            os.remove(os.path.join(root, name))
+      #os.rmdir("Rules")
+    #os.mkdir("Rules")
     f = open (improcess_bison_tab_cpp, "r")
     
     # delete current rules
